refactor(scan): node-free buildAuditReport via score-core leaf
buildAuditReport re-tainted the browser graph (@ast-grep/napi native binary +
node:fs/crypto/child_process) through audit-score/optimize/audit-verdict →
leaderboard.ts → scan.ts. leaderboard.ts eagerly imports scanPlugin (used only
in the CLI-only rankPlugins), so importing its pure scoring functions dragged
the whole native chain in.

Extract the pure scoring core into node-free leaf src/score-core.ts (PluginScore,
W_* weights, Deduction, gradeFor, reportDeductions, isEmptyMachine,
computeIntegrityScore, scoreReport — same pattern as core/ncd.ts / scan-core.ts).
leaderboard.ts re-exports them and keeps its node-only scanPlugin/rankPlugins;
audit-score/optimize/audit-verdict/audit-report import from score-core.

Now BOTH browser entry points are node-free: the dist require-graph union of
scan-files.js + audit-report.js has builtins = only node:zlib (→ pako in the
site build), pkgs = @iarna/toml/js-yaml/mvdan-sh. audit-report.js went 41 files
(+ ast-grep) → 6 files, zero node deps. Internal-only (no api drift);
coverage-neutral; parity firewall + full unit suite (100% cov) green.
